SEQNG-271 Split the security package moving the ui parts and activator to a sparate bundle #time 4h
