+++ bundle/edu.gemini.util.security.ext/src/main/resources/edu/gemini/util/security/auth/ui/icons/delete.gif
